Add support for Safari and always get the latest webdrivers
This version adds support for Safari on MacOS (no headless support)
and deals with some of its peculiarities (pre HTML objects not
containing newlines when accessed via getText).

Web driver versions are no longer hard coded and instead always
the latest one is fetched. If that's not the version installed on
the computer, you can adjust the settings temporarily in
browser_support/package.json.